Further refactor node loading
The main idea here is that the createContentDigest function from
Gatsby doesn’t have to be passed from sourceNodesArgs as a dependency.
According to the docs it just performs some hashing, so it has no
state and can be called directly. This further simplifies the
conversion from our datatypes to Gatsby nodes, since the conversion
is now just a plain function of T => NodeInput type.